Add files via upload
This is the scrolling model.
I could not delete the buttons without deleting the map. I couldn't figure out how to adapt the script.
I also could not figure out how to give the second map location the boundaries for the zoomed area. 
I would like to follow up later, but this is it for the class work, I'm afraid.